right x64

+++ checkers/Move.cpp
@@ -17,12 +17,12 @@ int operator==(const Move & m1, const Move & m2)
 	return 1;
 }
 
-void Move::SetVarEaten(int n1, ...) {
-	int * p = &n1;
+void Move::SetVarEaten(ull n1, ...) {
+	ull * p = &n1;
 	int n=0;
 	move = move&INVMASK_EATEN&INVMASK_NEATEN;
 	for (; (*p)&&(n<10); n++, p++) {
-		move = move|((long long)(*p) << (n * 4 + OFFSET_EATEN));
+		move = move|(*p << (n * 4 + OFFSET_EATEN));
 	}
 	SetNEaten(n);
 }
